Add an action to the DEAL button
What is a closure ?
A: A block of code that can be passed into a method call as a parameter.

What is a trailing closure ?
A: Its’ a closure that is taken out of the parameter list and moved to the end of the method call.

How do you handle the tap of a Button ?
A: By specifying a closure for the action parameter of the Button initialiser.